Add Hayvan Sağlığı Yönetim Sistemi Postman collection and enhance backend security configuration. Updated authentication and role management in controllers, payloads, and services. Improved CORS settings and JWT handling for better security and usability.

+++ backend/src/main/java/com/hayvansaglik/yonetim/payload/response/JwtResponse.java
@@ -10,12 +10,12 @@ import java.util.List;
 public class JwtResponse {
     private String token;
     private String type = "Bearer";
-    private Long id;
+    private Integer id;
     private String username;
     private String email;
     private List<String> roles;
 
-    public JwtResponse(String accessToken, Long id, String username, String email, List<String> roles) {
+    public JwtResponse(String accessToken, Integer id, String username, String email, List<String> roles) {
         this.token = accessToken;
         this.id = id;
         this.username = username;
